Fix OOB string access
The TextSystems seems to use a sort of compressed UTF-16 encoding,
whereas the zero low-byte on (7-bit) ASCII characters is omitted.
When accessing the next character of a string care must be taken to not
access a string past the NUL character, which is also encoded in a
single byte.